Split UnitTestStopWatch to a extendable class
